Add form for creating new contacts
Hooking into the select dropdown is not really possible, as we need to
create two different kinds of contact.

This patch adds a new form reachable from the top menu that can be used
to create new contacts. After the API returns success, the data is added
to the list of all contacts and also local cache is updated.

The layout of the page is split into a separate component that handles
displaying and hiding of the new dialog.

JIRA: PDC-2410